interdigitate pdfs
This script allows the user to scan the fronts and backs of papers and then merge them together in the correct order - note that usually scanning the back reverses the page order.  This script fixes that.  Good for doing 'double sided' scanning when you only have a single sided scanner to work with.